Move fetch_remote_resource to pymlstats/htmlparser
Leave pymlstats/utils only for local stuffs and try to group
remote stuffs narrowed to htmlparser and backend.
